feat(cli): continue searching until arriving at your home directory
a. If the option `--config` or a `FORMAT_PACKAGE_CONFIG` is provided, skip searching and attempt to load the configuration.

b. If the option `--config` or a `FORMAT_PACKAGE_CONFIG` is not provided, search for configurations until arriving at the home directory.

If no configuration is found, then the default configuration is used.